Move ScopeGAgent host AGUI mapper tests off endpoint reflection (cluster-010)
Refactor (iter5/cluster-010):
- Old pattern: cluster-003 moved draft-run streaming to the projection
  pipeline, but ScopeGAgentEndpoints still exposed endpoint-local
  TryMapEnvelopeToAguiEvent(EventEnvelope) and BuildToolApprovalStruct(Any)
  wrappers, and ScopeGAgentEndpointsTests reflected into those Host methods
  to assert raw EventEnvelope -> AGUIEvent mapping. Violated "Host 只做协议
  与组合" and locked the old API alive as test-locked residual.
- New principle: host endpoint tests verify protocol/interaction behavior
  only; mapper behavior is tested directly on ScopeGAgentAguiEventMapper.

Concrete changes:
- Deleted TryMapEnvelopeToAguiEvent and BuildToolApprovalStruct from
  ScopeGAgentEndpoints (no production callers remain).
- Created ScopeGAgentAguiEventMapperTests.cs calling the mapper directly
  (no reflection) for mapper + invalid-approval-payload cases.
- Added source-regression assertion in ScopeGAgentEndpointsTests.
